+++ 2SUM/utilities.h
@@ -2,9 +2,9 @@
 #define LINE_MAX_SIZE_CHAR 10000
 #define LINE_MAX_SIZE_INT 10000
 
-int printArray(int,int*, char*);
-int readArray(FILE*, int*, int**);
-int arraySwap(int**, int, int);
+int printArray(int,long long*, char*);
+int readArray(FILE*, int*, long long**);
+int arraySwap(long long**, int, int);
 int getNumbersFromString(char*, int *);
 int getLinesInFile(FILE*);
 int getNumFromFile(FILE*);
